Change Mission's FloatFields to DecimalFields so that table display looks nicer

+++ smdb/smdb/models.py
@@ -161,7 +161,9 @@ class Mission(models.Model):
     nav_track = models.LineStringField(
         srid=4326, spatial_index=True, blank=True, null=True
     )
-    track_length = models.FloatField(blank=True, null=True)
+    track_length = models.DecimalField(
+        max_digits=8, decimal_places=3, blank=True, null=True
+    )
     expedition = models.ForeignKey(
         Expedition, on_delete=models.CASCADE, blank=True, null=True
     )
@@ -175,7 +177,9 @@ class Mission(models.Model):
     vehicle_name = models.CharField(max_length=128, blank=True, null=True)
     start_date = models.DateTimeField(blank=True, null=True)
     end_date = models.DateTimeField(blank=True, null=True)
-    start_depth = models.FloatField(blank=True, null=True)
+    start_depth = models.DecimalField(
+        max_digits=8, decimal_places=3, blank=True, null=True
+    )
     start_point = models.PointField(
         srid=4326, spatial_index=True, dim=2, blank=True, null=True
     )
